int to long on cpp

+++ src/InferFunctionsSC.h
@@ -1,5 +1,5 @@
-void init_class(int _task_num, int _worker_num, double *_yDist, double *_pVec, double *_xVec);
+void init_class(long _task_num, long _worker_num, double *_yDist, double *_pVec, double *_xVec);
 
-double infer(int *pLabelMat);
+double infer(long *pLabelMat);
 
 void deconstruct();
